+++ RELEASE_NOTES.md
@@ -14,6 +14,12 @@ Breaking changes can still occur in minor versions and patches, though, if the a
 
 This version fixes the iOS 15 autorotate bug and adds a property that can suppress the `needsInputModeSwitchKey` warning.
 
+### ✨ New features
+
+* `KeyboardInputViewController` has a new `viewWillSetupKeyboard` that can be overridden to setup the keyboard at the proper time.
+* `KeyboardInputViewController` has a new `needsInputModeSwitchKeyOverride` that can be set to make an input controllers ignore the real value.
+* `KeyboardInputViewController` has a new, static `needsInputModeSwitchKeyOverride` that can be set to make all input controllers ignore the real value.
+
 ### 🗑 Deprecations
 
 * `KeyboardInputViewController` `setup(with:)` stack view variant is deprecated and will be removed in 5.0.